backend/requirements.txt
# AegisNode Backend Dependencies
# ─────────────────────────────────
# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PDF Processing
PyMuPDF==1.25.3

# Text Processing
langchain-text-splitters==0.3.5

# Embeddings
sentence-transformers==3.4.1

# Vector Store
faiss-cpu==1.9.0.post1

# LLM Client
httpx==0.28.1

# Database
sqlalchemy==2.0.37
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
